Moving player and coordinator to package to allow testing
