+++ solutions/05-battleship/run-05.py
@@ -6,41 +6,32 @@ import multiprocessing
 import os
 import random
 from concurrent.futures import ProcessPoolExecutor
-from typing import ClassVar
 from typing import Literal
 
 from academy.agent import action
-from academy.agent import Agent
-from academy.agent import loop
 from academy.exchange.cloud import HttpExchangeFactory
-from academy.handle import Handle
 from academy.logging import init_logging
 from academy.manager import Manager
 from globus_compute_sdk import Executor as GCExecutor
 
 from academy_tutorial.battleship import Board
 from academy_tutorial.battleship import Crd
-from academy_tutorial.battleship import Game
+from academy_tutorial.coordinator import Coordinator
+from academy_tutorial.player import BattleshipPlayer
 
 EXCHANGE_ADDRESS = 'https://exchange.academy-agents.org'
 logger = logging.getLogger(__name__)
 
 
-class BattleshipPlayer(Agent):
+class MyBattleshipPlayer(BattleshipPlayer):
     def __init__(
         self,
     ) -> None:
-        from academy_tutorial.battleship import Board
-
         super().__init__()
         self.guesses = Board()
 
     @action
     async def get_move(self) -> Crd:
-        import asyncio
-
-        from academy_tutorial.battleship import Crd
-
         await asyncio.sleep(1)
         while True:
             row = random.randint(0, self.guesses.size - 1)
@@ -65,73 +56,11 @@ class BattleshipPlayer(Agent):
 
     @action
     async def new_game(self, ships: list[int], size: int = 10) -> Board:
-        from academy_tutorial.battleship import Board
-        from academy_tutorial.battleship import Crd
-
         self.guesses = Board(size)
         my_board = Board(size)
         for i, ship in enumerate(ships):
             my_board.place_ship(Crd(i, 0), ship, 'horizontal')
         return my_board
-
-
-class Coordinator(Agent):
-    _default_ships: ClassVar[list[int]] = [5, 5, 4, 3, 2]
-
-    def __init__(
-        self,
-        player_0: Handle[BattleshipPlayer],
-        player_1: Handle[BattleshipPlayer],
-        *,
-        size: int = 10,
-        ships: list[int] | None = None,
-    ) -> None:
-        from academy_tutorial.battleship import Board
-        from academy_tutorial.battleship import Game
-
-        super().__init__()
-        self.player_0 = player_0
-        self.player_1 = player_1
-        self.game_state = Game(Board(), Board())
-        self.ships = ships or self._default_ships
-        self.stats = [0, 0]
-
-    async def game(self, shutdown: asyncio.Event) -> int:
-        while not shutdown.is_set():
-            attack = await self.player_0.get_move()
-            result = self.game_state.attack(0, attack)
-            await self.player_0.notify_result(attack, result)
-            await self.player_1.notify_move(attack)
-            if self.game_state.check_winner() >= 0:
-                return self.game_state.check_winner()
-
-            attack = await self.player_1.get_move()
-            result = self.game_state.attack(1, attack)
-            await self.player_1.notify_result(attack, result)
-            await self.player_0.notify_move(attack)
-            if self.game_state.check_winner() >= 0:
-                return self.game_state.check_winner()
-
-        return -1
-
-    @loop
-    async def play_games(self, shutdown: asyncio.Event) -> None:
-        from academy_tutorial.battleship import Game
-
-        while not shutdown.is_set():
-            player_0_board = await self.player_0.new_game(self.ships)
-            player_1_board = await self.player_1.new_game(self.ships)
-            self.game_state = Game(player_0_board, player_1_board)
-            winner = await self.game(shutdown)
-            self.stats[winner] += 1
-
-    @action
-    async def get_game_state(self) -> Game | None:
-        return self.game_state
-
-    @action
-    async def get_player_stats(self) -> list[int]:
-        return self.stats
 
 
 async def main() -> int:
@@ -156,11 +85,8 @@ async def main() -> int:
         # process pool executor.
         executors=executor,
     ) as manager:
-        # Launch each of the three agents, each implementing a different
-        # behavior. The returned type is a handle to that agent used to
-        # invoke actions.
-        player_1 = await manager.launch(BattleshipPlayer)
-        player_2 = await manager.launch(BattleshipPlayer)
+        player_1 = await manager.launch(MyBattleshipPlayer)
+        player_2 = await manager.launch(MyBattleshipPlayer)
         coordinator = await manager.launch(
             Coordinator,
             args=(player_1, player_2),
@@ -187,9 +113,6 @@ async def main() -> int:
             else:
                 print('Unknown command')
             print('-----------------------------------------------------')
-
-        # Upon exit, the Manager context will instruct each agent to shutdown,
-        # closing their respective handles, and shutting down the executors.
 
     return 0
